Added comment/post deletes, updated CSS
Added delete button to posts and comments made by currently logged-in user. Updated database logic (ON DELETE CASCADE) for necessary foreign keys. Uses stored procedures for deletion process.

Updated CSS for login and post pages. An issue that likely won't be solved in time is that the pages are not responsive and mostly break upon resizing.