Implemented API refresh token functionality
API access token, refresh token, and access token expiration date in epoch time are now stored in the database.

When the access token expires, the Strava API is called with the refresh token parameter to get a new user access token and expiration date that will be updated in the database.

Added a /likedposts page that displays all posts a user has liked. It can be accessed through a user's profile or the sidebar.

Added a comments count to all posts that is updated when a comment is added or deleted under a post.

Made minor changes to CSS.