User activities now displayed on /activities page
Data for all activities stored in the database from the Strava API is now displayed for the user on the /activities page.

Created a function for converting time to the correct format for display on the  /activities page.

Added a "Refresh Activities" button to the /activities page that redirects to the /getactivities route, updating the database with any new data from the Strava API.

Updated create_db script to include the new 'getactivities' procedure.