Activities can now be linked when making posts
Added a dropdown menu to select an activity from in the post form of the home page, allowing a user to link an activity to their new post.

Updated logic of make_post procedure to insert the activity_id into the new post row, and the new post's ID into the respective activity row.